# app/domain/request/admin.go
package request

type UpsertTemplate struct {
	Code             string                       `json:"code" binding:"required"`
	TextTh           string                       `json:"textTh" binding:"required"`
	TextEn           string                       `json:"textEn" binding:"required"`
	ChannelOverrides map[string]TemplateOverrides `json:"channelOverrides"`
	Active           bool                         `json:"active"`
}

type TemplateOverrides struct {
	TextTh string `json:"textTh"`
	TextEn string `json:"textEn"`
}

type UpdateBranchSetting struct {
	ShopName           string `json:"shopName"`
	RetentionHours     int    `json:"retentionHours" binding:"required"`
	CooldownSeconds    int    `json:"cooldownSeconds" binding:"required"`
	ConfirmMethod      string `json:"confirmMethod" binding:"required"`
	SmsCreditThreshold int    `json:"smsCreditThreshold"`
	ContactChannel     string `json:"contactChannel"`
}

type SetPin struct {
	Pin           string `json:"pin" binding:"required,len=6,numeric"`
	ConfirmMethod string `json:"confirmMethod" binding:"required"`
}

type UpsertPermission struct {
	UserId            string   `json:"userId" binding:"required"`
	BranchId          string   `json:"branchId" binding:"required"`
	Phone             string   `json:"phone"`
	AllowedEventTypes []string `json:"allowedEventTypes"`
	IsTestRecipient   bool     `json:"isTestRecipient"`
	Active            bool     `json:"active"`
}

type CreateQr struct {
	BranchId string `json:"branchId" binding:"required"`
	TableNo  string `json:"tableNo"`
}